Inline needed bits of ipv4_apps into arp.lua
